Style Cop CleanUp
